Move backend functions to submodule.
